+++ src/components/Dialog/Dialog.styles.ts
@@ -6,7 +6,6 @@ export interface DialogVariables {
   body: ViewStyle;
   container: ViewStyle;
   modalContainer: ViewStyle;
-  overlay: ViewStyle & { cursor?: string };
 }
 
 export const getDialogVariables = (theme: Theme): DialogVariables => {
@@ -30,24 +29,12 @@ export const getDialogVariables = (theme: Theme): DialogVariables => {
       justifyContent: 'center',
       width: '100%',
     },
-    overlay: {
-      backgroundColor: theme.colors.background.overlay,
-      bottom: 0,
-      cursor: 'auto',
-      height: '100%',
-      left: 0,
-      position: 'absolute',
-      right: 0,
-      top: 0,
-      width: '100%',
-    },
   };
 };
 
 export interface DialogStyles {
   bodyStyle: ViewStyle;
   containerStyle: ViewStyle;
-  overlayStyle: ViewStyle;
   modalContainerStyle: ViewStyle;
 }
 export type GetDialogStyles = (theme: Theme) => DialogStyles;
@@ -59,6 +46,5 @@ export const getDialogStyles: GetDialogStyles = theme => {
     bodyStyle: dialogVariables.body,
     containerStyle: dialogVariables.container,
     modalContainerStyle: dialogVariables.modalContainer,
-    overlayStyle: dialogVariables.overlay,
   };
 };