+++ src/components/Positioner/Positioner.styles.ts
@@ -5,7 +5,6 @@ import { Theme } from '../../theme/ThemeInterface';
 export interface PositionerVariables {
   positioner: ViewStyle;
   modalContainer: ViewStyle;
-  overlay: ViewStyle & { cursor?: string };
 }
 
 export const getPositionerVariables = (theme: Theme): PositionerVariables => {
@@ -17,17 +16,6 @@ export const getPositionerVariables = (theme: Theme): PositionerVariables => {
       justifyContent: 'center',
       width: '100%',
     },
-    overlay: {
-      backgroundColor: 'transparent',
-      bottom: 0,
-      cursor: 'auto',
-      height: '100%',
-      left: 0,
-      position: 'absolute',
-      right: 0,
-      top: 0,
-      width: '100%',
-    },
     positioner: {
       position: 'absolute',
       zIndex: 1,
@@ -37,7 +25,6 @@ export const getPositionerVariables = (theme: Theme): PositionerVariables => {
 
 export interface PositionerStyles {
   positionerStyle: ViewStyle;
-  overlayStyle: ViewStyle;
   modalContainerStyle: ViewStyle;
 }
 
@@ -48,7 +35,6 @@ export const getPositionerStyles: GetPositionerStyles = theme => {
 
   return {
     modalContainerStyle: positionerVariables.modalContainer,
-    overlayStyle: positionerVariables.overlay,
     positionerStyle: positionerVariables.positioner,
   };
 };